LayoutTests/platform/chromium/inspector/debugger/function-details-expected.txt

Tests that Debugger.getFunctionDetails command returns correct location. Bug 71808

Debugger was enabled.

Running: testGetFirstLineFunctionDetails
firstLineFunction type = function
Function details: 
lineNumber: 5
columnNumber: 36
scriptId is valid: true
name: firstLineFunction
displayName: undefined
inferredName: undefined
scopeChain #0: global; <global object properties omitted>

Running: testGetNonFirstLineFunctionDetails
notFirstLineFunction type = function
Function details: 
lineNumber: 10
columnNumber: 32
scriptId is valid: true
name: notFirstLineFunction
displayName: undefined
inferredName: undefined
scopeChain #0: global; <global object properties omitted>

Running: testGetDetailsOfFunctionWithInferredName
obj.m type = function
Function details: 
lineNumber: 16
columnNumber: 15
scriptId is valid: true
name: undefined
displayName: undefined
inferredName: obj.m
scopeChain #0: global; <global object properties omitted>

Running: testGetDetailsOfFunctionWithDisplayName
functionWithDisplayName type = function
Function details: 
lineNumber: 19
columnNumber: 32
scriptId is valid: true
name: functionWithDisplayName
displayName: undefined
inferredName: undefined
scopeChain #0: global; <global object properties omitted>

Running: testSmallClosure
smallClosure type = function
Function details: 
lineNumber: 22
columnNumber: 49
scriptId is valid: true
name: undefined
displayName: undefined
inferredName: undefined
scopeChain #0: closure; p: "Capybara"
scopeChain #1: global; <global object properties omitted>

Running: testBigClosure
bigClosure type = function
Function details: 
lineNumber: 35
columnNumber: 27
scriptId is valid: true
name: undefined
displayName: undefined
inferredName: undefined
scopeChain #0: catch; ex: <no string representation>
scopeChain #1: with; __proto__: <no string representation>,e: 7,f: 5,u: <no value>,v: <no value>
scopeChain #2: closure; arguments: <no string representation>,o: <no string representation>,p: <no string representation>
scopeChain #3: global; <global object properties omitted>
Debugger was disabled.
